requirements.txt
asgiref==3.11.1
Django==5.2.11
django-filter==25.2
djangorestframework==3.16.1
pillow==12.1.1
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
